Rewriting back-matter export.
